backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
redis==5.0.8
stripe==10.12.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0


